BJN-24: Organized socket code further into handlers and fixed a bug in url generator
