+++ Backend/BlackJackNow/src/services/gameTableService.ts
@@ -14,4 +14,3 @@ export const createGameTable = async (context: AppContext) => {
     throw error;
   }
 };
-
