+++ package.json
@@ -28,7 +28,8 @@
     "ci:frontend": "cd Frontend/BlackJackNow && npm run ci",
     "ci:backend": "cd Backend/BlackJackNow && npm run ci",
     "ci:test": "docker compose -f docker-compose.test.yml up --build --exit-code-from backend-test",
-    "start:docker": "docker compose up --build --watch",
+    "start:docker": "docker compose up --build -d",
+    "start:docker:watch": "docker compose up --build --watch",
     "start:docker:nomiddleware": "DISABLE_MIDDLEWARE=true npm run start:docker",
     "stop:docker": "docker compose stop",
     "clean:docker": "docker image prune --force",
